+++ ChangeLog.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+### Fixed
+
+- fixed repository methods with `Example` using `StringMatcher.CONTAINING` (issue #113)
+
 ## [3.0.0-RC.1] - 2018-08-13
 
 ### Added